refactor: Separate auth logic into separate service files
