+++ server/internal/services/validateToken_service.go
@@ -1,0 +1,5 @@
+package services
+
+func (s *authService) Authenticate(tokenString string) error {
+	return s.tokenGenerator.VerifyToken(tokenString)
+}
